requirements.txt
numpy==2.1.1
pillow==10.4.0
pycryptodome==3.20.0
